refactor(runtime): put resources together
