+++ packages/runtime/src/aurelia.ts
@@ -1,7 +1,7 @@
 import { DI, IContainer, IRegistry, PLATFORM, Registration } from '@aurelia/kernel';
 import { INode } from './dom.interfaces';
 import { LifecycleFlags } from './observation';
-import { CustomElementResource, ICustomElement, ICustomElementType } from './templating/custom-element';
+import { CustomElementResource, ICustomElement, ICustomElementType } from './resources/custom-element';
 import { IRenderingEngine } from './templating/lifecycle-render';
 
 export interface ISinglePageApp {
