+++ packages/runtime/src/binding/ast.ts
@@ -1,12 +1,12 @@
 import { IIndexable, IServiceLocator, PLATFORM, Reporter, StrictPrimitive } from '@aurelia/kernel';
 import { IBindScope } from '../lifecycle';
 import { Collection, IBindingContext, IOverrideContext, IScope, LifecycleFlags, ObservedCollection } from '../observation';
+import { BindingBehaviorResource } from '../resources/binding-behavior';
+import { IValueConverter, ValueConverterResource } from '../resources/value-converter';
 import { IBinding } from './binding';
-import { BindingBehaviorResource } from './binding-behavior';
 import { BindingContext } from './binding-context';
 import { IConnectableBinding } from './connectable';
 import { ISignaler } from './signaler';
-import { IValueConverter, ValueConverterResource } from './value-converter';
 
 export type IsPrimary = AccessThis | AccessScope | ArrayLiteral | ObjectLiteral | PrimitiveLiteral | Template;
 export type IsLiteral = ArrayLiteral | ObjectLiteral | PrimitiveLiteral | Template;
