+++ packages/runtime/src/html-renderer.ts
@@ -13,8 +13,8 @@ import { customAttributeKey, customElementKey, ICallBindingInstruction, IHydrate
 import { DOM } from './dom';
 import { IElement, IHTMLElement, INode, IRenderLocation } from './dom.interfaces';
 import { IAttach, IAttachables, IBindables, IBindScope, IRenderable, IRenderContext } from './lifecycle';
-import { ICustomAttribute } from './templating/custom-attribute';
-import { ICustomElement } from './templating/custom-element';
+import { ICustomAttribute } from './resources/custom-attribute';
+import { ICustomElement } from './resources/custom-element';
 import { IElementHydrationOptions, IInstructionRenderer, instructionRenderer, IRenderer, IRenderingEngine } from './templating/lifecycle-render';
 
 const slice = Array.prototype.slice;
