+++ packages/runtime/src/resources/binding-behaviors/self-binding-behavior.ts
@@ -1,9 +1,9 @@
 import { IRegistry, Reporter } from '@aurelia/kernel';
+import { findOriginalEventTarget } from '../../binding/event-manager';
+import { Listener } from '../../binding/listener';
 import { IEvent, INode } from '../../dom.interfaces';
 import { IScope, LifecycleFlags } from '../../observation';
 import { bindingBehavior } from '../binding-behavior';
-import { findOriginalEventTarget } from '../event-manager';
-import { Listener } from '../listener';
 
 /** @internal */
 export function handleSelfEvent(this: SelfableBinding, event: IEvent): ReturnType<Listener['callSource']> {
