+++ packages/runtime/src/resources/custom-attribute.ts
@@ -3,9 +3,9 @@ import { BindingMode } from '../binding/binding-mode';
 import { customAttributeKey, customAttributeName, IAttributeDefinition } from '../definitions';
 import { Hooks, IAttach, IBindScope, ILifecycleHooks, ILifecycleUnbindAfterDetach, IRenderable } from '../lifecycle';
 import { IChangeTracker } from '../observation';
-import { $attachAttribute, $cacheAttribute, $detachAttribute } from './lifecycle-attach';
-import { $bindAttribute, $unbindAttribute } from './lifecycle-bind';
-import { $hydrateAttribute, IRenderingEngine } from './lifecycle-render';
+import { $attachAttribute, $cacheAttribute, $detachAttribute } from '../templating/lifecycle-attach';
+import { $bindAttribute, $unbindAttribute } from '../templating/lifecycle-bind';
+import { $hydrateAttribute, IRenderingEngine } from '../templating/lifecycle-render';
 
 type CustomAttributeStaticProperties = Pick<Immutable<Required<IAttributeDefinition>>, 'bindables'>;
