+++ packages/runtime/src/resources/custom-attributes/if.ts
@@ -2,7 +2,7 @@ import { inject, IRegistry } from '@aurelia/kernel';
 import { IRenderLocation } from '../../dom.interfaces';
 import { CompositionCoordinator, IView, IViewFactory } from '../../lifecycle';
 import { LifecycleFlags } from '../../observation';
-import { bindable } from '../bindable';
+import { bindable } from '../../templating/bindable';
 import { ICustomAttribute, templateController } from '../custom-attribute';
 
 export interface If extends ICustomAttribute {}
