+++ packages/runtime/src/resources/custom-attributes/with.ts
@@ -3,7 +3,7 @@ import { Scope } from '../../binding/binding-context';
 import { IRenderLocation } from '../../dom.interfaces';
 import { IBindScope, IView, IViewFactory, State } from '../../lifecycle';
 import { IBindingContext, LifecycleFlags } from '../../observation';
-import { bindable } from '../bindable';
+import { bindable } from '../../templating/bindable';
 import { ICustomAttribute, templateController } from '../custom-attribute';
 
 export interface With extends ICustomAttribute {}
