+++ packages/runtime/src/templating/create-element.ts
@@ -3,7 +3,7 @@ import { buildTemplateDefinition, isTargetedInstruction, TargetedInstruction, Ta
 import { DOM } from '../dom';
 import { INode } from '../dom.interfaces';
 import { IRenderContext, IView, IViewFactory } from '../lifecycle';
-import { ICustomElementType } from './custom-element';
+import { ICustomElementType } from '../resources/custom-element';
 import { IRenderingEngine, ITemplate } from './lifecycle-render';
 
 const slice = Array.prototype.slice;
