+++ packages/runtime/src/templating/lifecycle-attach.ts
@@ -2,8 +2,8 @@ import { Tracer, Writable } from '@aurelia/kernel';
 import { INode } from '../dom.interfaces';
 import { Hooks, IView, State } from '../lifecycle';
 import { LifecycleFlags } from '../observation';
-import { ICustomAttribute } from './custom-attribute';
-import { ICustomElement } from './custom-element';
+import { ICustomAttribute } from '../resources/custom-attribute';
+import { ICustomElement } from '../resources/custom-element';
 
 const slice = Array.prototype.slice;
